* 'master' of https://github.com/AY2324S1-CS2103T-T14-1/tp: (33 commits)
  Fix typo mistake
  Correct small typo
  Update DG - Edit Command description
  Fix typo mistake
  Update sequence diagram
  Fix a line that was too long
  Fix indentation
  Resolve merge conflict
  Update prefix summary
  Modify increment sequence diagram
  Update method name
  Add a new line at EOF
  Update DG - Edit Command
  Update command summary - import/export
  Update command summary
  Improve clarity of UG History Command
  Change hard coded maximum salary
  Update the DG for Find Command
  Add prefix summary
  Add Table of contents
  ...

# Conflicts:
#	docs/UserGuide.md

+++ src/test/java/seedu/address/model/person/SalaryWithinRangePredicateTest.java
@@ -63,8 +63,8 @@ class SalaryWithinRangePredicateTest {
 
     @Test
     public void toStringMethod() {
-        int keywordLowerBound = 1000;
-        int keywordUpperBound = 5000;
+        double keywordLowerBound = 1000.0;
+        double keywordUpperBound = 5000.0;
         SalaryWithinRangePredicate predicate = new SalaryWithinRangePredicate(keywordLowerBound, keywordUpperBound);
 
         String expected = SalaryWithinRangePredicate.class.getCanonicalName() + "{lowerBound=" + keywordLowerBound
